+++ crates/kubernation-core/src/k8s/mod.rs
@@ -1,4 +1,5 @@
 pub mod actions;
+pub mod browse;
 pub mod client;
 pub mod logs;
 pub mod metrics;